+++ CHANGELOG.md
@@ -21,7 +21,8 @@ Change Log
 16. All views now starts with 'window-resized' coop by default if view.coop is not set to false;
 17. Added _.isjQueryObject() util function;
 18. Added _.deepClone*() util functions;
-19. Upgraded to use BB 1.3, jQuery 2.2, rapheal 2.2 and flexlayout 0.3;
+19. Aligned 'view:navigate-to' to be after 'view:ready' instead of 'region:show';
+20. Upgraded to use BB 1.3, jQuery 2.2, rapheal 2.2 and flexlayout 0.3;
 
 working on ...